+++ questionnaire/questionnaire/utils/taskRunner/tasks/transformAndUpload/index.test.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const {transformQuestionnaire, getIsFatal, updateCaseReferenceWithYear} = require('.');
-const questionnaireFixture = require('../../questionnaire/test-fixtures/res/questionnaireCompleteForCheckYourAnswers');
-const questionnaire = require('../../questionnaire/questionnaire/questionnaire');
+const {transformQuestionnaire} = require('.');
+const questionnaireFixture = require('../../../../../test-fixtures/res/questionnaireCompleteForCheckYourAnswers');
+const questionnaire = require('../../../../questionnaire');
 
 describe('Integration Service', () => {
     // Arrange
@@ -61,19 +61,5 @@ describe('Integration Service', () => {
                 return question.id === 'q--new-or-existing-application';
             });
         expect(newOrExistingQuestion.hideOnSummary).toBeTruthy();
-    });
-
-    it('Should get false for isFatal if not fatal', () => {
-        const isFatal = getIsFatal(questionnaireFixture);
-        expect(isFatal).toBeFalsy();
-    });
-
-    it('Should format the case reference correctly', () => {
-        const referenceNumber = '123456';
-        const dateModified = new Date(2022, 1, 1);
-
-        const caseReference = updateCaseReferenceWithYear(referenceNumber, dateModified);
-
-        expect(caseReference).toBe('22\\123456');
     });
 });
